Refactor auth context, remove unused dependencies, and clean up imports
