+++ package.json
@@ -14,20 +14,10 @@
     "@reduxjs/toolkit": "^2.8.2",
     "@tanstack/react-query": "^5.85.3",
     "axios": "^1.11.0",
-    "bcrypt": "^6.0.0",
     "chart.js": "^4.5.0",
-    "cookie-parser": "^1.4.7",
-    "cors": "^2.8.5",
-    "dotenv": "^17.2.1",
-    "express": "^5.1.0",
-    "express-validator": "^7.2.1",
-    "init": "^0.1.2",
-    "jsonwebtoken": "^9.0.2",
     "jwt-decode": "^3.1.2",
     "lucide-react": "^0.539.0",
-    "mongoose": "^8.17.2",
-    "multer": "^2.0.2",
-    "node": "^24.6.0",
+
     "react": "^19.1.1",
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^19.1.1",
