+++ src/context/useAuth.js
@@ -1,0 +1,11 @@
+// useAuth.js - Custom hook for authentication
+import { useContext } from "react";
+import { AuthContext } from "./AuthContext.jsx";
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error("useAuth must be used within an AuthProvider");
+  }
+  return context;
+};
